open_telemetory/otel-getting-started/pyproject.toml
[tool.poetry]
name = "otel-getting-started"
version = "0.1.0"
description = ""
authors = ["Naoya Yoshimura <yoshimura708x@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
flask = "^3.0.3"
opentelemetry-distro = "^0.46b0"
opentelemetry-exporter-otlp = "^1.25.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
